+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 - Support to global cluster
 - Support for monitoring features
 
+## [0.0.2] - 2023-11-30
+
+- Add support for storage encryption.
+- Add support for deletion protection.
+- Add support for auto-version upgrades. 
+
 
 ## [0.0.1] - 2023-11-09